Refactoring: Rename Method/Class ChangeChefDataService
The naming of the class ChangeChefDataService as well as the method
ChangeData was not clear enough. The new names are more descriptive
and should make it easier to understand what the class and method do.

Also extracted two methods from the original ChangeData method as they
fulfilled two smaller single responsibilities. The original ChangeData
method (now ChangeChefNameAndImage) delegates the two new methods.

Co-authored-by: Fabian Wolf <fabianwlf@outlook.de>